Refactor src/api to dismantle barrel file
- Create src/api/constants.ts for API constants
- Create src/api/factories.ts for endpoint factories
- Create src/api/client.ts for main API implementation
- Update src/api/types.ts with API types
- Update src/api/user.ts to fix import
- Update src/api/index.ts to be a pure re-export
- Update PREVENT_CONFLICTS.md
